make twig extension lazy by using a Runtime
