+++ composer.json
@@ -131,7 +131,7 @@
     },
 
     "conflict": {
-        "twig/twig":                    "<1.34"
+        "twig/twig":                    "<1.35|>=2.0,<2.4.4"
     },
 
     "scripts": {
